fix target time
